+++ metrics_sdk/lib/opentelemetry/sdk/metrics/meter.rb
@@ -11,7 +11,7 @@ module OpenTelemetry
     module Metrics
       # {Meter} is the SDK implementation of {OpenTelemetry::Metrics::Meter}.
       class Meter < OpenTelemetry::Metrics::Meter
-        NAME_REGEX = /\A[a-zA-Z][-.\w]{0,62}\z/
+        NAME_REGEX = %r{\A[a-zA-Z][-./\w]{0,254}\z}
 
         # @api private
         #
@@ -26,6 +26,30 @@ module OpenTelemetry
           @instrument_registry = {}
           @instrumentation_scope = InstrumentationScope.new(name, version)
           @meter_provider = meter_provider
+        end
+
+        # Multiple-instrument callbacks
+        # Callbacks registered after the time of instrument creation MAY be associated with multiple instruments.
+        # Related spec: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#multiple-instrument-callbacks
+        # Related spec: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#synchronous-instrument-api
+        #
+        # @param [Array] instruments A list (or tuple, etc.) of Instruments used in the callback function.
+        # @param [Proc] callback A callback function
+        #
+        # It is RECOMMENDED that the API authors use one of the following forms for the callback function:
+        # The list (or tuple, etc.) returned by the callback function contains (Instrument, Measurement) pairs.
+        # the Observable Result parameter receives an additional (Instrument, Measurement) pairs
+        # Here it chose the second form
+        def register_callback(instruments, callback)
+          instruments.each do |instrument|
+            instrument.register_callback(callback)
+          end
+        end
+
+        def unregister(instruments, callback)
+          instruments.each do |instrument|
+            instrument.unregister(callback)
+          end
         end
 
         # @api private