+++ metrics_sdk/test/opentelemetry/sdk/metrics/instrument/gauge_test.rb
@@ -28,7 +28,6 @@ describe OpenTelemetry::SDK::Metrics::Instrument::Gauge do
     _(last_snapshot[0].instrumentation_scope.name).must_equal('test')
     _(last_snapshot[0].data_points[0].attributes).must_equal('foo' => 'bar')
     _(last_snapshot[0].data_points[0].value).must_equal(-2)
-    _(last_snapshot[0].aggregation_temporality).must_equal(:delta)
   end
 
   it 'gauge should count 1 for last recording' do