+++ logs_sdk/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History: opentelemetry-logs-sdk
 
+### v0.3.0 / 2025-08-14
+
+- ADDED: Update error message for invalid log record attribute
+- ADDED: Set logs exporter default to OTLP
+
 ### v0.2.0 / 2025-02-25
 
 - ADDED: Support 3.1 Min Version